React Table. Int

+++ buildSrc/src/main/kotlin/karakum/table/Converter.kt
@@ -64,7 +64,7 @@ private fun convertConst(
     val body = source.replace(": {", " {")
         .replace("\n    ", "\n    val ")
         .replace("<any>", "<*>")
-        .replace("number", "Double")
+        .replace("number", "Int")
 
     val content = "external $type $body"
     return ConversionResult(name, content)
